refactor: aggregate wal configs
