+++ src/cmd/Cargo.toml
@@ -33,6 +33,7 @@ common-telemetry = { workspace = true, features = [
     "deadlock_detection",
 ] }
 common-time.workspace = true
+common-wal.workspace = true
 config = "0.13"
 datanode.workspace = true
 datatypes.workspace = true
