+++ src/cmd/src/datanode.rs
@@ -18,8 +18,8 @@ use std::time::Duration;
 use async_trait::async_trait;
 use catalog::kvbackend::MetaKvBackend;
 use clap::Parser;
-use common_config::WalConfig;
 use common_telemetry::{info, logging};
+use common_wal::config::DatanodeWalConfig as WalConfig;
 use datanode::config::DatanodeOptions;
 use datanode::datanode::{Datanode, DatanodeBuilder};
 use datanode::service::DatanodeServiceBuilder;
