+++ src/common/meta/src/error.rs
@@ -14,7 +14,7 @@
 
 use std::str::Utf8Error;
 
-use common_config::wal::WalOptions;
+use common_config::wal_options::WalOptions;
 use common_error::ext::{BoxedError, ErrorExt};
 use common_error::status_code::StatusCode;
 use common_macro::stack_trace_debug;
