+++ src/common/wal/Cargo.toml
@@ -1,16 +1,24 @@
 [package]
-name = "common-config"
+name = "common-wal"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 
-# TODO(niebayes): move content in common-wal into this package.
+[features]
+testing = []
 
 [dependencies]
 common-base.workspace = true
 humantime-serde.workspace = true
+rskafka.workspace = true
 serde.workspace = true
 serde_with.workspace = true
+# TODO(niebayes): move to dev dependencies.
+common-telemetry.workspace = true
+futures-util.workspace = true
+
+# TODO(niebayes): maybe move wal options into this package.
 
 [dev-dependencies]
 serde_json.workspace = true
+toml.workspace = true
