+++ src/datanode/Cargo.toml
@@ -35,6 +35,7 @@ common-recordbatch.workspace = true
 common-runtime.workspace = true
 common-telemetry.workspace = true
 common-time.workspace = true
+common-wal.workspace = true
 dashmap.workspace = true
 datafusion-common.workspace = true
 datafusion-expr.workspace = true
