+++ src/meta-srv/src/metasrv.rs
@@ -26,12 +26,12 @@ use common_meta::key::TableMetadataManagerRef;
 use common_meta::kv_backend::{KvBackendRef, ResettableKvBackend, ResettableKvBackendRef};
 use common_meta::peer::Peer;
 use common_meta::region_keeper::MemoryRegionKeeperRef;
-use common_meta::wal::options_allocator::WalOptionsAllocatorRef;
-use common_meta::wal::WalConfig;
+use common_meta::wal_options_allocator::WalOptionsAllocatorRef;
 use common_procedure::options::ProcedureConfig;
 use common_procedure::ProcedureManagerRef;
 use common_telemetry::logging::LoggingOptions;
 use common_telemetry::{error, info, warn};
+use common_wal::config::MetasrvWalConfig as WalConfig;
 use serde::{Deserialize, Serialize};
 use servers::export_metrics::ExportMetricsOption;
 use servers::http::HttpOptions;
