+++ src/mito2/Cargo.toml
@@ -55,7 +55,7 @@ puffin.workspace = true
 regex = "1.5"
 serde = { version = "1.0", features = ["derive"] }
 serde_json.workspace = true
-serde_with = "3"
+serde_with.workspace = true
 smallvec.workspace = true
 snafu.workspace = true
 store-api.workspace = true
