+++ src/mito2/src/region.rs
@@ -22,7 +22,7 @@ use std::collections::HashMap;
 use std::sync::atomic::{AtomicBool, AtomicI64, Ordering};
 use std::sync::{Arc, RwLock};
 
-use common_config::wal::WalOptions;
+use common_config::wal_options::WalOptions;
 use common_telemetry::info;
 use common_time::util::current_time_millis;
 use snafu::{ensure, OptionExt};
