+++ src/mito2/src/region_write_ctx.rs
@@ -16,7 +16,7 @@ use std::mem;
 use std::sync::Arc;
 
 use api::v1::{Mutation, OpType, Rows, WalEntry};
-use common_config::wal::WalOptions;
+use common_config::wal_options::WalOptions;
 use snafu::ResultExt;
 use store_api::logstore::LogStore;
 use store_api::storage::{RegionId, SequenceNumber};
