+++ tests-integration/Cargo.toml
@@ -30,6 +30,7 @@ common-recordbatch.workspace = true
 common-runtime.workspace = true
 common-telemetry.workspace = true
 common-test-util.workspace = true
+common-wal.workspace = true
 datanode = { workspace = true }
 datatypes.workspace = true
 dotenv = "0.15"
